requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.5
django-celery-beat==2.3.0
django-celery-results==2.3.1
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.2.1
eventlet==0.33.1
Flask==2.1.0
gevent==21.12.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mysqlclient==2.1.1
nulltype==2.3.1
oauthlib==3.2.0
packaging==21.3
plaid-python==9.1.0
prompt-toolkit==3.0.29
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyngrok==5.1.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
structlog==21.5.0
toml==0.10.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
